% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isobxr_solvers.R
\name{ana_slvr}
\alias{ana_slvr}
\title{Analytically solve stable isotope box models}
\usage{
ana_slvr(input_path)
}
\arguments{
\item{input_path}{path to the INPUT file containing all commands for the run.
\cr (file name structure: RUN name + _INPUT.xlsx)}
}
\value{
The function returns the analytically determined evolution of stable
isotope compositions in all boxes over the run duration as specified in
INPUT file. The function outputs are as follows:
\enumerate{
\item OUT data file with initial and final size and delta values in each boxes.
\cr (file name structure: RUN name + _A_1_OUT.csv)
\item ODE_SOLNs data file summarizing outputs of the analytical solutions of the ODES
(eigenvalues, eigenvectors, relaxation times, constants according to initial conditions).
\cr (file name structure: RUN name + _A_2_ODE_SOLNs.csv)
\item evD data file of the evolution with time of the delta values in each boxes.
\cr (file name structure: RUN name + _A_3_evD.csv)
}
}
\description{
An analytical solver of the system of ordinary differential
equations (ODES) of stable isotope ratios of element X in all boxes.
The analytical solver finds the eigenvalues and eigenvectors of the ODES.
Given the initial conditions as specified in INPUT.xlsx file, it determines the
set of analytical solutions that describes the evolution of isotope ratios
in each box over time.
}
