% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{calculation_gauge}
\alias{calculation_gauge}
\title{Gauge of calculation}
\usage{
calculation_gauge(i, len)
}
\arguments{
\item{i}{Calculating step number}

\item{len}{Total calculation steps}
}
\value{
A print summarizing the progress of calculation
}
\description{
Prints a gauge of the calculation progress
}
\keyword{internal}
