% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_FORCINGS}
\alias{check_FORCINGS}
\title{check_FORCINGS}
\usage{
check_FORCINGS(type, FORCING, isobxr_master)
}
\arguments{
\item{type}{Type of forcing. As character. \cr
One value among: "RAYLEIGH", "ALPHA", "SIZE", "DELTA"}

\item{FORCING}{Forcing data frame.}

\item{isobxr_master}{Isobxr master as a list as returned from read_isobxr_master}
}
\value{
aborts run if a mistake is found
}
\description{
A function to check the formatting data frames used as forcing inputs \cr
in several functions (e.g., sim.single_run)
}
\keyword{internal}
