% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{clear_subset}
\alias{clear_subset}
\title{Clear a subset}
\usage{
clear_subset(dataset)
}
\arguments{
\item{dataset}{Previously subsetted data frame}
}
\value{
The cleared data frame.
}
\description{
Takes a previously subsetted data frame and
clears the lost levels of factor type columns and resets the rown numbering.
}
