% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_composite_model.R
\name{compose_isobxr}
\alias{compose_isobxr}
\title{Compose a stable isotope box model scenario}
\usage{
compose_isobxr(
  workdir,
  SERIES_ID,
  time_units,
  COMPO_MASTER,
  plot_HIDE_BOXES_delta,
  plot_HIDE_BOXES_size,
  EACH_RUN_DIGEST = FALSE,
  to_CPS_DIGEST_CSVs = FALSE
)
}
\arguments{
\item{workdir}{Working directory in which the general master file (0_ISOBXR_MASTER.xlsx)
and composite master file (e.g., 0_COMPO_MASTER.xlsx)
are found and where the output files will be stored.
\cr (string of characters).}

\item{SERIES_ID}{Name of the composite model series the run belongs to, that will determine
the folder in which the output files of this composite run
will be stored. A composite run number will be automatically linked to it and no run will
overwrite a previous composite run.
\cr (string of characters).}

\item{time_units}{Vector defining the initial time units (identical to time unit
used for fluxes) followed by the time unit used for the graphic output.
\cr (vector of two strings of characters, eg. c("days", "years")).
\cr should be picked amongst: "micros" "ms"     "s"      "min"    "h"      "d"      "wk"     "mo"     "yr"     "kyr"    "Myr"    "Gyr"}

\item{COMPO_MASTER}{Name of the composite master file (e.g., 0_COMPO_MASTER.xlsx),
defining the composite run scenario.
\cr (string of characters).}

\item{plot_HIDE_BOXES_delta}{Vector of strings of characters defining the names of the
boxes to hide in the plot of the delta values as a function of time edited as a pdf.
\cr (e.g., c("BOX_A", "BOX_C"))}

\item{plot_HIDE_BOXES_size}{Vector of strings of characters defining the names of the
boxes to hide in the plot of the box sizes (masses of X) as a function of time edited
as a pdf.}

\item{EACH_RUN_DIGEST}{Get each full digest for each model run (INPUT xlsx, diagrams, plot and csv outputs). Default is FALSE.}

\item{to_CPS_DIGEST_CSVs}{Export all global csv outputs (full evD and full evS). Default is FALSE.}
}
\value{
If non existing, the fonction creates and stores all outputs
in a dedicated composite SERIES directory located in working directory.
\cr Directory name structure: 3_CPS + SERIES_ID + YYY, where YYY is an automically set composite run number between 001 and 999.
\cr No overwriting of previous composite runs is possible.
\enumerate{
\item Creates the set of inputs and outputs for each successive \emph{n} runs, numbered from to 1 to \emph{n} in an XXXX format (see \code{\link{run_isobxr}} documentation).
\cr Named with the following format: CPS + SERIES_ID + YYY + XXXX
\item Archives the composite master file.
\cr (file name structure: 0_CPS + SERIES_ID + YYY + COMPO_MASTER.xlsx)
\item Archives the local LOG for the given composite run.
\cr (file name structure: 0_CPS + SERIES_ID + YYY + LOG.csv)
\item Stores the evolution of delta values with time in all boxes over the \emph{n} runs that constitute the composite run scenario.
\cr (file name structure: 0_CPS + SERIES_ID + YYY + evD.csv)
\item Stores the evolution of box sizes (masses of X) with time in all boxes over the \emph{n} runs that constitute the composite run scenario.
\cr (file name structure: 0_CPS + SERIES_ID + YYY + evS.csv)
\item Edits the pdf of the all in one plot of the evolution of delta values + sizes in all non hidden boxes.
\cr (file name structure: 0_CPS + SERIES_ID + YYY + p_evDS.pdf)
\item Edits the pdf of multiple plots of the evolution of delta values in all non hidden boxes.
\cr (file name structure: 0_CPS + SERIES_ID + YYY + pf_evD.pdf)
\item Edits the pdf of multiple plots of the evolution of box sizes in all non hidden boxes.
\cr (file name structure: 0_CPS + SERIES_ID + YYY + pf_evS.pdf)
}
}
\description{
A function to compose an isobxr box model scenario,
defined by a series of \emph{n} model runs inheriting from previous run's final state and
with the possibility to force conditions at each run, namely:
\enumerate{
\item fluxes (all or overwriting only a subset of them)
\item isotope fractionation coefficients (all or overwriting only a subset of them)
\item box sizes
\item rayleigh distillation
\item isotope composition of a source box
\item (...)
}
}
\seealso{
Documentation on \code{\link{run_isobxr}}
}
