% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{corr_stats}
\alias{corr_stats}
\title{corr_stats}
\usage{
corr_stats(data_lm, X, Y, as_character = F, n = 3, m = 4)
}
\arguments{
\item{data_lm}{a data frame with X and Y data}

\item{X}{name of X variable column (character)}

\item{Y}{name of Y variable column (character)}

\item{as_character}{reports statistics formatted as characters. Default is FALSE.}

\item{n}{number of digits to display in correlation coefficient/rho. Default is 3. Only used if as_character = T.}

\item{m}{number of digits to display in regression equation parameters (slope, y-intercept). Default is 4. Only used if as_character = T.}
}
\value{
report of regression and correlation tests
}
\description{
edits dataframe report of linear regression and correlation tests (Pearson and Spearman)
}
\keyword{internal}
