% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{dec_0}
\alias{dec_0}
\title{Print a number with 0 decimal figures}
\usage{
dec_0(x)
}
\arguments{
\item{x}{Numerical value}
}
\value{
A print (string of character) with 0 decimal figure.
}
\description{
Takes a numerical value and returns a print with 0 decimal figures.
}
