% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{dec_1}
\alias{dec_1}
\title{Print a number with 1 decimal figures}
\usage{
dec_1(x)
}
\arguments{
\item{x}{Numerical value}
}
\value{
A character string with 1 decimal figure.
}
\description{
Takes a numerical value and returns a print with 1 decimal figures.
}
\keyword{internal}
