% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{dec_2}
\alias{dec_2}
\title{Print a number with 2 decimal figures}
\usage{
dec_2(x)
}
\arguments{
\item{x}{Numerical value}
}
\value{
A character string with 2 decimal figures.
}
\description{
Takes a numerical value and returns a print with 2 decimal figures.
}
\keyword{internal}
