% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{dec_3}
\alias{dec_3}
\title{Print a number with 3 decimal figures}
\usage{
dec_3(x)
}
\arguments{
\item{x}{Numerical value}
}
\value{
A character string with 3 decimal figures.
}
\description{
Takes a numerical value and returns a print with 3 decimal figures.
}
\keyword{internal}
