% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{dec_4}
\alias{dec_4}
\title{Print a number with 4 decimal figures}
\usage{
dec_4(x)
}
\arguments{
\item{x}{Numerical value}
}
\value{
A character string with 4 decimal figures.
}
\description{
Takes a numerical value and returns a print with 4 decimal figures.
}
\keyword{internal}
