% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{dec_n}
\alias{dec_n}
\title{Print a number with n decimal figures}
\usage{
dec_n(x, n)
}
\arguments{
\item{x}{Numerical value}

\item{n}{number of decimals}
}
\value{
A character string with 4 decimal figures.
}
\description{
Takes a numerical value and returns a print with n decimal figures.
}
\keyword{internal}
