% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{del_NaN_rows}
\alias{del_NaN_rows}
\title{Delete rows of a data frame containing NaN values in a given column}
\usage{
del_NaN_rows(dataframe, by_col, resetrows)
}
\arguments{
\item{dataframe}{Data frame}

\item{by_col}{Name of column from which NaN values should be removed}

\item{resetrows}{Logical value to reset the row numbering or not.}
}
\value{
Data frame without the rows in which there previously were NaN values in column by_col
}
\description{
Delete rows of a data frame containing NaN values in a given column
}
