% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{get_portability_data}
\alias{get_portability_data}
\title{Collect local portability data (for DEV only)}
\usage{
get_portability_data(workdir)
}
\arguments{
\item{workdir}{directory where Rdata local summary will be exported}
}
\value{
a Rdata file with loc wd, version, sessionInfo, installed packages
}
\description{
collect local platform/session/versions data for portability (for DEV only)
}
\keyword{internal}
