% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isobxr-package.R
\docType{package}
\name{isobxr-package}
\alias{isobxr}
\alias{isobxr-package}
\title{isobxr: Stable Isotope Box Modelling in R}
\description{
A set of functions to run simple and composite box-models to describe the dynamic or static distribution of stable isotopes in open or closed systems. The package also allows the sweeping of many parameters in both static and dynamic conditions. The mathematical models used in this package are derived from Albarede, 1995, Introduction to Geochemical Modelling, Cambridge University Press, Cambridge \doi{10.1017/CBO9780511622960}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ttacail/isobxr}
  \item \url{https://ttacail.github.io/isobxr_web/}
  \item \url{https://ttacail.github.io/isobxr/}
  \item Report bugs at \url{https://github.com/ttacail/isobxr/issues}
}

}
\author{
\strong{Maintainer}: Theo Tacail \email{theo.tacail@protonmail.com}

Authors:
\itemize{
  \item Theo Tacail \email{theo.tacail@protonmail.com}
}

}
\keyword{internal}
