% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isobxr_solvers.R
\name{num_slvr}
\alias{num_slvr}
\title{Numerically solve stable isotope box models}
\usage{
num_slvr(input_path)
}
\arguments{
\item{input_path}{path to the INPUT file containing all commands for the run.
\cr (file name structure: RUN name + _INPUT.xlsx)}
}
\value{
The function returns the numerically determined evolution of stable
isotope compositions and mass of element X in all boxes over the run duration as
specified in INPUT file. The function outputs are as follows:
\enumerate{
\item OUT data file storing initial and final size and delta values in all boxes.
\cr (file name structure: RUN name + _N_1_OUT.csv)
\item evS data file storing the evolution with time of the sizes (masses of element X) of all boxes.
\cr (file name structure: RUN name + _N_2_evS.csv)
\item evD data file storing the evolution with time of the delta values in all boxes.
\cr (file name structure: RUN name + _N_3_evD.csv)
}
}
\description{
A numerical solver of the system of ordinary differential
equations (ODES) of stable isotope ratios of element X in all boxes of a system.
The numerical solver uses the ode function of the deSolve package
to integrate the stable isotopes ratios in each box. It allows the
calculation of the evolution of stable isotope ratio even in the case of
unbalanced outward and inward fluxes of element X in a given box
resulting in the accumulation or loss of element X.
}
