% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fit_final_space.R
\name{plot_correlogram_all_CIfit}
\alias{plot_correlogram_all_CIfit}
\title{plot_correlogram_all_CIfit}
\usage{
plot_correlogram_all_CIfit(DF, names.swp.ADSR)
}
\arguments{
\item{DF}{fit.final_space data frame}

\item{names.swp.ADSR}{names of swept ADSR parameters (alpha, delta, size, rayleigh, custom expressions), as vector}
}
\value{
A correlogram of CI fitted ADSR parameter values
}
\description{
Plot correlogram of CI fitted ADSR parameter values
}
\keyword{internal}
