% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{plot_diagram}
\alias{plot_diagram}
\title{plot_diagram}
\usage{
plot_diagram(input, title, matrix_layout, BOXES_master_loc, COEFF_FLUX)
}
\arguments{
\item{input}{matrix of fluxes or coefficients}

\item{title}{title of diagram}

\item{matrix_layout}{matrix of X,Y layout}

\item{BOXES_master_loc}{BOXES_master_loc}

\item{COEFF_FLUX}{"COEFF" or "FLUX"}
}
\value{
diagrams for console or pdf edition (not used in shinobxr)
}
\description{
plot diagrams for fluxes and coefficients using the qgraph network function
}
\keyword{internal}
