% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fit_final_space.R
\name{plot_freq_flux}
\alias{plot_freq_flux}
\title{plot_freq_flux}
\usage{
plot_freq_flux(DF, parameter_subsets)
}
\arguments{
\item{DF}{fit.final_space data frame}

\item{parameter_subsets}{fitted flux lists subsets passed from fit.final_space}
}
\value{
A frequency plot of fitted parameter values.
}
\description{
Plot frequency profiles by fitted flux lists
}
\keyword{internal}
