% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fit_final_space.R
\name{plot_lda}
\alias{plot_lda}
\title{plot_lda}
\usage{
plot_lda(DF, names.swp.ADSR)
}
\arguments{
\item{DF}{fit.final_space data frame}

\item{names.swp.ADSR}{names of swept ADSR parameters (alpha, delta, size, rayleigh, custom expressions), as vector}
}
\value{
A linear discriminant analysis by quartiles of summed squared residuals for all ADSR parameters
}
\description{
Plot linear discriminant analysis by quartiles of summed squared residuals for all ADSR parameters
}
\keyword{internal}
