% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fit_final_space.R
\name{plot_LS_surfaces_lists}
\alias{plot_LS_surfaces_lists}
\title{plot_LS_surfaces_lists}
\usage{
plot_LS_surfaces_lists(DF, names.swp.lists)
}
\arguments{
\item{DF}{fit.final_space data frame}

\item{names.swp.lists}{names of swept list parameters (flux and coeff lists), as vector}
}
\value{
A plot of surfaces of summed squared residuals for flux lists and coeff lists.
}
\description{
Plot surfaces of summed squared residuals.
}
\keyword{internal}
