% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fit_final_space.R
\name{plot_sim_distrib}
\alias{plot_sim_distrib}
\title{plot_sim_distrib}
\usage{
plot_sim_distrib(DF, fit.counts, observations, bx.fit)
}
\arguments{
\item{DF}{fit.final_space data frame}

\item{fit.counts}{counts of simulations by groups passed from fit.final_space}

\item{observations}{data frame of observations}

\item{bx.fit}{list of boxes fitted and unfitted groups and subgroups passed from fit.final_space}
}
\value{
A plot of distributions of all and fitted simulations
}
\description{
Plot distributions of all and fitted simulations
}
\keyword{internal}
