% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fit_final_space.R
\name{plot_sim_obs}
\alias{plot_sim_obs}
\title{plot_sim_obs}
\usage{
plot_sim_obs(DF, bx.fit, fit.counts)
}
\arguments{
\item{DF}{fit.final_space data frame}

\item{bx.fit}{list of boxes fitted and unfitted groups and subgroups passed from fit.final_space}

\item{fit.counts}{counts of simulations by groups passed from fit.final_space}
}
\value{
A plot of fitted simulated delta values against observations
}
\description{
Plot simulations against observations
}
\keyword{internal}
