% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_isobxr}
\alias{run_isobxr}
\title{Run isobxr stable isotope box model}
\usage{
run_isobxr(
  workdir,
  SERIES_ID,
  flux_list_name,
  coeff_list_name,
  t_lim,
  nb_steps,
  time_units,
  FORCING_RAYLEIGH = NULL,
  FORCING_SIZE = NULL,
  FORCING_DELTA = NULL,
  FORCING_ALPHA = NULL,
  COMPOSITE = FALSE,
  COMPO_SERIES_n = NaN,
  COMPO_SERIES_FAMILY = NaN,
  EXPLORER = FALSE,
  EXPLO_SERIES_n = NaN,
  EXPLO_SERIES_FAMILY = NaN,
  HIDE_PRINTS = FALSE,
  PLOT_DIAGRAMS = TRUE,
  PLOT_evD = FALSE
)
}
\arguments{
\item{workdir}{Working directory in which the master file (0_ISOBXR_MASTER.xlsx)
is found and where the output files will be stored.
\cr (string of characters)}

\item{SERIES_ID}{Name of the model series the run belongs to, that will determine
the folder in which the output files will be stored.
\cr (string of characters)}

\item{flux_list_name}{Name of the list of fluxes to be used for the run as
defined in a single column of the FLUXES sheet of the 0_ISOBXR_MASTER.xlsx file.
\cr (string of characters)}

\item{coeff_list_name}{Name of the list of fractionation coefficients to be used
for the run as defined in a single column of the COEFFS sheet of the
0_ISOBXR_MASTER.xlsx file.
\cr (string of characters)}

\item{t_lim}{Run duration, given in the same time units as the fluxes.
\cr (integer)}

\item{nb_steps}{Number of steps that will define the resolution of the run.
\cr (integer)}

\item{time_units}{Vector defining the initial time units (identical to time unit
used for fluxes) followed by the time unit used for the graphic output.
\cr (vector of two strings of characters, eg. c("days", "years"))}

\item{FORCING_RAYLEIGH}{\emph{OPTIONAL} Dataframe describing the forcing on a Rayleigh
distillation driven fractionation coefficient as a function of flux intensities
and a fundamental fractionation coefficient (Dataframe formating details in
read me).
\cr Default is NULL.}

\item{FORCING_SIZE}{\emph{OPTIONAL} Dataframe describing the forcing on one or several box
sizes (mass of element X). The newly defined sizes for the given set of boxes
shown in dataframe will overwrite the sizes as defined in 0_ISOBXR_MASTER.xlsx file.
\cr Default is NULL.}

\item{FORCING_DELTA}{\emph{OPTIONAL} Dataframe describing the forcing on one or several boxes
initial isotope composition described as delta (Dataframe formating details in read me).
The newly defined delta values for the given set of boxes shown in dataframe will
overwrite the delta values as defined in ISOBXR_MASTER file.
\cr Default is NULL.}

\item{FORCING_ALPHA}{\emph{OPTIONAL} Dataframe describing the forcing on one or several
fractionation coefficients linked to fluxes from one reservoir to another
(Dataframe formating details in read me). The newly defined alpha values
for the given set of boxes shown in dataframe will overwrite the alpha values
as defined in ISOBXR_MASTER file.
\cr Default is NULL.}

\item{COMPOSITE}{\emph{NOT TO BE USED IN SINGLE RUN} Logical value automatically defined in \code{\link{compose_isobxr}}.
\cr Default is FALSE.}

\item{COMPO_SERIES_n}{\emph{NOT TO BE USED IN SINGLE RUN} Iteration of the composite run for the given series it belongs to,
automatically defined in \code{\link{compose_isobxr}}.
\cr Default is NaN.}

\item{COMPO_SERIES_FAMILY}{\emph{NOT TO BE USED IN SINGLE RUN} Composite run series family, automatically defined in
\code{\link{compose_isobxr}}.
\cr Default is NaN.}

\item{EXPLORER}{\emph{NOT TO BE USED IN SINGLE RUN} Logical value automatically defined in \code{\link{sweep_steady}}
or \code{\link{sweep_dyn}}.
\cr Default is FALSE.}

\item{EXPLO_SERIES_n}{\emph{NOT TO BE USED IN SINGLE RUN} Iteration of the sweep run for the given series it belongs to,
automatically defined in \code{\link{sweep_steady}} or \code{\link{sweep_dyn}}.
\cr Default is NaN.}

\item{EXPLO_SERIES_FAMILY}{\emph{NOT TO BE USED IN SINGLE RUN} Sweep run series family, automatically defined in
\code{\link{sweep_steady}} or \code{\link{sweep_dyn}}.
\cr Default is NaN.}

\item{HIDE_PRINTS}{\emph{OPTIONAL} Logical value determining whether to run outputs details in R console.
This parameter will not hide the warnings regarding the automatic update of
the run duration in case of the emptying of a box.
\cr Default is FALSE.}

\item{PLOT_DIAGRAMS}{\emph{OPTIONAL} Logical value to edit pdf of box model diagram or not.
\cr Default is TRUE.}

\item{PLOT_evD}{\emph{OPTIONAL} Logical value to edit pdf of delta time evolution plot or not.
\cr Default is TRUE.}
}
\value{
If function is run independently and if the directory is non existing,
the fonction creates and stores all outputs
in a SERIES directory located in working directory.
\cr Directory name structure: 2_RUN + SERIES_ID

\enumerate{
\item Automatically sets a XXXX run number between 0001 and 9999. The outputs do not overwrite possible identical previously performed runs.

\item Creates an INPUT file with all run conditions, located in SERIES directory.
\cr (file name structure: SERIES_ID + XXXX + _INPUT.xlsx)

\item Edits a Box model diagram of flux (DIAGf pdf) of element X (mass per time unit)
between all boxes. (Optional, depends on parameter PLOT_DIAGRAM), located in SERIES directory.
\cr (file name structure: SERIES_ID + XXXX + DIAGf + _flux_list_name.pdf)

\item Edits a Box model diagram of isotope fractionation coefficients (alphas, DIAGa pdf)
between all boxes (optional, depends on parameter PLOT_DIAGRAM), located in SERIES directory.
\cr (file name structure: SERIES_ID + XXXX + DIAGf + _flux_list_name.pdf)

\item Creates a subdirectory for numerical outputs.
\cr (directory name structure: SERIES_ID + XXXX + _OUT)

\enumerate{
\item Stores \code{\link{num_slvr}} or \code{\link{ana_slvr}} outputs in subdirectory.

\item Edits a pdf plot of the time dependent evolution of delta values, with a logarithmic x axis time scale (optional).
\cr (file name structure: SERIES_ID + XXXX + _plot_evD.pdf)
}

\item Creates or updates the general log file, located in general working directory.
\cr (file name: 1_LOG.csv)
}
}
\description{
A function to run the isobxr stable isotope box model,
assessing the design of the model and automatically running \code{\link{num_slvr}}
or \code{\link{ana_slvr}} depending on the conditions.
}
\seealso{
Documentation on \code{\link{num_slvr}} or \code{\link{ana_slvr}} functions.
}
