% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyPlots.R
\name{shinobxr_app}
\alias{shinobxr_app}
\title{Call isobxr plot shiny app}
\usage{
shinobxr_app()
}
\value{
No return value, called for launch of the html based shiny app.
}
\description{
A function to call the isobxr plot shiny app to interactively plot outputs from \code{\link{compose_isobxr}},
\code{\link{sweep_steady}} and \code{\link{sweep_dyn}}. \cr
The function takes no arguments but requires the definition of a working directory where all SERIES directory are stored. \cr
This working directory needs to be defined as a character string and stored in a variable called workdir.
For instance: \cr
workdir = "User/isobxr_working_directory"
}
