% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{significance_pval_df}
\alias{significance_pval_df}
\title{edits a full dataframe of significance symbols from a dataframe of p-values}
\usage{
significance_pval_df(x)
}
\arguments{
\item{x}{a data frame column of p-values or a pvalues as double (alone or in vector).}
}
\value{
p-values symbols as dataframe or as string.
}
\description{
edits a full dataframe of significance symbols from a dataframe of p-values
\cr Symbols as asterisks denote p-values lower than .0001, .001, .01, .05. ns refers to non significant (higher than .5).
}
\keyword{internal}
