% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyPlots.R
\name{silence_shiny_deps}
\alias{silence_shiny_deps}
\title{silently use packages called in shiny app / silence the warning "Namespaces in Imports field not imported from"}
\usage{
silence_shiny_deps(param)
}
\arguments{
\item{param}{param}
}
\value{
No return value, called for side effects.
}
\description{
collect local platform/session/versions data for portability (for DEV only)
}
\keyword{internal}
