% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sweepers.R
\name{sweep_dyn}
\alias{sweep_dyn}
\title{Sweep the space of two parameters during a dynamic run}
\usage{
sweep_dyn(
  workdir,
  SERIES_ID,
  time_units,
  EXPLO_MASTER,
  EXPLO_AXIS_1,
  EXPLO_AXIS_2,
  to_DYN_DIGEST_CSVs = FALSE
)
}
\arguments{
\item{workdir}{Working directory in which the general master file (0_ISOBXR_MASTER.xlsx)
and dynamic sweep master file (e.g., 0_EXPLO_DYN_MASTER.xlsx)
are found and where the output files will be stored.
\cr (string of characters)}

\item{SERIES_ID}{Name of the sweep series the run belongs to, that will determine
the folder in which the output files of this sweep run will be stored.
A sweep run number in YYY format will be automatically linked to it and no run will
overwrite a previous sweep run.
\cr (string of characters)}

\item{time_units}{Vector defining the initial time units (identical to time unit
used for fluxes) followed by the time unit used for the graphic output.
\cr (vector of two strings of characters, eg. c("days", "years"))}

\item{EXPLO_MASTER}{Name of the dynamic sweep master file (e.g., 0_EXPLO_DYN_MASTER.xlsx),
defining the dynamic sweep run scenario.
\cr (string of characters)}

\item{EXPLO_AXIS_1}{Set of values of sweeping parameter 1}

\item{EXPLO_AXIS_2}{Set of values of sweeping parameter 2}

\item{to_DYN_DIGEST_CSVs}{Export all global csv outputs (full evD and full evS) or not to the DYN DIGEST folder.
\cr Logical value, default is FALSE.}
}
\value{
Creates and stores all outputs in a SERIES directory
located in working directory, automatically numbered.
\cr Directory name structure: 4_DYN + SERIES_ID + YYY, YYY being an automically set dynamic sweep run number between 001 and 999.
\cr No overwriting of previous dynamic sweep runs is possible.
\enumerate{
\item Calculates the number of runs the sweeping will require depending on the sweeped parameters.
\item Asks the user confirmation to run sweep_dyn,
as the run calculation time depends on the number of successive sweeping runs.
\item Creates  all input (IN.RDA) and output (OUT.RDA) files for all successive \emph{n} runs, numbered from to 1 to \emph{n} in an XXXX format (see \code{\link{run_isobxr}} documentation).
\item Summarizes results in the DIGEST folder as follows :
\enumerate{
\item Archive of local LOG file (LOG.csv) for all runs of the steady sweeping.
\item Archive of steady sweep master file (.xlsx).
\item Evolution of delta values (evD.Rds) with time in all boxes over the \emph{n} runs that constitute the steady sweep run.
\item Evolution of box sizes (masses of element X, evS.Rds) with time in all boxes over the \emph{n} runs that constitute the steady sweep run.
}
}
}
\description{
A function to assess the influence of two parameters (varying
over a range of values) on dynamic evolution of given model.
}
