% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sweepers.R
\name{sweep_steady}
\alias{sweep_steady}
\title{Sweep the space of two parameters at the final state of a run}
\usage{
sweep_steady(
  workdir,
  SERIES_ID,
  time_units,
  EXPLO_MASTER,
  EXPLO_AXIS_1,
  EXPLO_AXIS_2,
  to_STD_DIGEST_CSVs = FALSE
)
}
\arguments{
\item{workdir}{Working directory in which the general master file (0_ISOBXR_MASTER.xlsx)
and steady sweep master file (e.g., 0_EXPLO_STEADY_MASTER.xlsx)
are found and where the output files will be stored.
\cr (string of characters)}

\item{SERIES_ID}{Name of the sweep series the run belongs to, that will determine
the folder in which the output files of this sweep run will be stored.
A sweep run number in YYY format will be automatically linked to it and no run will
overwrite a previous sweep run.
\cr (string of characters)}

\item{time_units}{Vector defining the initial time units (identical to time unit
used for fluxes) followed by the time unit used for the graphic output.
\cr (vector of two strings of characters, eg. c("days", "years"))}

\item{EXPLO_MASTER}{Name of the steady sweep master file (e.g., 0_EXPLO_STEADY_MASTER.xlsx),
defining the steady sweep run scenario.
\cr (string of characters)}

\item{EXPLO_AXIS_1}{Set of values of sweeping parameter 1}

\item{EXPLO_AXIS_2}{Set of values of sweeping parameter 2}

\item{to_STD_DIGEST_CSVs}{Export all global csv outputs (full evD and full evS) or not to the STD DIGEST folder.
\cr Logical value, default is FALSE.}
}
\value{
The fonction creates and stores all outputs in a SERIES directory
located in working directory, automatically numbered.
\cr Directory name structure: 4_STD + SERIES_ID + YYY, YYY being an automically set steady sweep run number between 001 and 999.
\cr No overwriting of previous steady sweep runs is possible.
\enumerate{
\item Calculates the number of runs the sweeping will require depending on the sweeped parameters.
\item Asks the user confirmation to run sweep_steady,
as the run calculation time depends on the number of successive sweeping runs.
\item Creates the set of inputs and outputs for the single initial run only (see \code{\link{run_isobxr}} documentation).
\item Summarizes results in the DIGEST folder as follows :
\enumerate{
\item Archive of local LOG file (LOG.csv) for all runs of the steady sweeping.
\item Archive of steady sweep master file (.xlsx).
\item Evolution of delta values (evD.Rds) with time in all boxes over the \emph{n} runs that constitute the steady sweep run.
\item Evolution of box sizes (masses of element X, evS.Rds) with time in all boxes over the \emph{n} runs that constitute the steady sweep run
\item Final state of delta values (evD_final.Rds) with time in all boxes over the \emph{n} runs that constitute the steady sweep run.
\item Final state of box sizes (masses of element X, evS_final.Rds) with time in all boxes over the \emph{n} runs that constitute the steady sweep run.
}
}
}
\description{
A function to assess the influence of two parameters (varying
over a range of values) on the final state of a given model.
}
