% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{tempdir_tree}
\alias{tempdir_tree}
\title{tempdir_tree}
\usage{
tempdir_tree()
}
\value{
a print of the directory tree
}
\description{
display tree of the temporary isobxr directory
}
\keyword{internal}
