% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{time_converter}
\alias{time_converter}
\title{convert time units in a data frame column}
\usage{
time_converter(
  dataframe,
  time_colname,
  conv_timecolname,
  former_unit,
  new_unit
)
}
\arguments{
\item{dataframe}{data frame for which a column with numerical time values should be converted to another unit}

\item{time_colname}{column with time values that need to be converted to a different time unit}

\item{conv_timecolname}{name of the column once time units converted. Can be identical to time_colname}

\item{former_unit}{former time unit (amongst: "micros" "ms"     "s"      "min"    "h"      "d"      "wk"     "mo"     "yr"     "kyr"    "Myr"    "Gyr")}

\item{new_unit}{new time unit (amongst: "micros" "ms"     "s"      "min"    "h"      "d"      "wk"     "mo"     "yr"     "kyr"    "Myr"    "Gyr")}
}
\description{
Convert the time units in plots
}
