% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{to_tmpdir}
\alias{to_tmpdir}
\title{address_to_tmpdir}
\usage{
to_tmpdir(dir_or_file)
}
\arguments{
\item{dir_or_file}{directory or file to address in temporary directory}
}
\value{
path to temporary subdir/file
}
\description{
address_to_tmpdir
}
\keyword{internal}
