% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{using_extdata_tutorial}
\alias{using_extdata_tutorial}
\title{using_extdata_tutorial}
\usage{
using_extdata_tutorial(workdir, save_run_outputs, plot_results)
}
\arguments{
\item{workdir}{working directory value refering to extdata tutorial files, several options among the following: \cr
c("/Users/username/Documents/1_ABC_tutorial",
"use_isobxr_demonstration_files",
system.file("extdata", package = "isobxr"),
system.file("extdata", "0_ISOBXR_MASTER.xlsx", package = "isobxr"))}
}
\value{
a vector with a boolean stating whether this is a tutorial mode, a workdir value allowing reading of extdata tutorial files, save_run_outputs = FALSE, plot_results = TRUE
}
\description{
identifies workdir value refering to use of extdata tutorial files, returns correct workdir value, \cr
prevents saving outputs locally (save_run_outputs = FALSE) \cr
forces display of default graphical output to R session (plot_results = TRUE)
}
\keyword{internal}
