% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{clear_subset}
\alias{clear_subset}
\title{Clear a subset}
\usage{
clear_subset(dataset)
}
\arguments{
\item{dataset}{Previously subsetted dataframe.}
}
\value{
Cleared dataframe.
}
\description{
Takes a previously subsetted dataframe, clears its deleted levels and resets its row indexes.
}
\keyword{internal}
