% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_composite_model.R
\name{compose_isobxr}
\alias{compose_isobxr}
\title{Compose a stable isotope box model scenario}
\usage{
compose_isobxr(
  workdir,
  SERIES_ID,
  time_units,
  COMPO_MASTER,
  plot_HIDE_BOXES_delta = NULL,
  plot_HIDE_BOXES_size = NULL,
  EACH_RUN_DIGEST = FALSE,
  to_CPS_DIGEST_CSVs = FALSE
)
}
\arguments{
\item{workdir}{Working directory of \strong{\emph{0_ISOBXR_MASTER.xlsx}} master file, \cr
of the composite master file (e.g., \strong{\emph{0_COMPO_MASTER.xlsx}}) \cr
and where output files will be stored. \cr
(character string)}

\item{SERIES_ID}{Name of the composite model series the run belongs to. \cr
It determines the folder in which the output files will be stored for this composite run.\cr
A composite run number is automatically linked to it,
subsequent runs can not overwrite a previous composite run.\cr
(character string)}

\item{time_units}{Vector defining the initial time unit
(identical to unit used in fluxes), \cr
followed by the time unit used for the graphical output.\cr
Character string, to be selected  amongst the following:\cr
\emph{micros, ms, s, min, h, d, wk, mo, yr, kyr, Myr, Gyr}\cr
e.g.,  c("d", "yr") to convert days into years}

\item{COMPO_MASTER}{Name of the composite master file (e.g., \strong{\emph{0_COMPO_MASTER.xlsx}}),
defining the composite run scenario. \cr
(character string)}

\item{plot_HIDE_BOXES_delta}{\emph{OPTIONAL} \cr
Vector of character strings, \cr
defining the names of the boxes to hide in the plot of the delta values as a function of time, edited as a pdf.
\cr (e.g., c("BOX_A", "BOX_C"))
\cr Default is NULL (no box hidden).}

\item{plot_HIDE_BOXES_size}{\emph{OPTIONAL} \cr
Vector of character strings, \cr
defining the names of the boxes to hide in the plot of the box sizes (masses of X) as a function of time, edited as a pdf.
\cr (e.g., c("BOX_A", "BOX_C"))
\cr Default is NULL (no box hidden).}

\item{EACH_RUN_DIGEST}{\emph{OPTIONAL} \cr
Logical value. \cr
Edits full digests for each model run
(all optional outputs of \code{\link{run_isobxr}} function) if TRUE. \cr
Default is FALSE.}

\item{to_CPS_DIGEST_CSVs}{\emph{OPTIONAL} \cr
Logical value. \cr
Exports all global csv outputs to \strong{\emph{0_CPS_DIGEST}} folder (full evD and full evS) if TRUE. \cr
Default is FALSE.}
}
\value{
Creates and stores all outputs in a dedicated composite SERIES directory located in working directory,
with the following name structure: \cr
\strong{\emph{3_CPS + SERIES_ID + YYY}}, where YYY is a composite scenario number automically set between 001 and 999. \cr
No overwriting of previous composite runs is possible.

\enumerate{
\item Creates the set of inputs and outputs for all successive \emph{n} runs,\cr
numbered from to 1 to \emph{n} in an XXXX format with the following format: \cr
\strong{\emph{CPS + SERIES_ID + YYY + XXXX + IN.Rda}} \cr
\strong{\emph{CPS + SERIES_ID + YYY + XXXX + OUT.Rda}} \cr
(see \code{\link{run_isobxr}} documentation)

\item Writes summarized results in the  \strong{\emph{0_CPS_DIGEST}} folder:
\enumerate{
\item LOG file of local composite run. \cr
(file name structure:  \strong{\emph{CPS + SERIES_ID + YYY + _LOG.csv}})
\item Composite master file. \cr
(file name structure:  \strong{\emph{CPS + SERIES_ID + YYY + _MASTER.xlsx}})
\item Dataset of temporal evolution of delta values (evD) in all boxes over the \emph{n} runs that constitute the composite run scenario \cr
(file name structure: \strong{\emph{CPS + SERIES_ID + YYY + evD.RDS}})
\item Dataset of temporal evolution of box sizes (evS, masses of X) in all boxes over the \emph{n} runs that constitute the composite run scenario \cr
(file name structure: \strong{\emph{CPS + SERIES_ID + YYY + evS.RDS}})
\item All-in-one plot of the evolution of delta values + sizes in all non hidden boxes. \cr
(file name structure: \strong{\emph{CPS + SERIES_ID + YYY + p_evDS.pdf}})
\item Multiple plots of the evolution of delta values in all non hidden boxes. \cr
(file name structure: \strong{\emph{CPS + SERIES_ID + YYY + pf_evD.pdf}})
\item Multiple plots of the evolution of box sizes in all non hidden boxes. \cr
(file name structure: \strong{\emph{CPS + SERIES_ID + YYY + pf_evS.pdf}})
}
}
}
\description{
A function to compose an isobxr box model scenario,
defined by a series of \emph{n} successive runs, \cr
each run inheriting from the final state conditions of the previous run. \cr
It is possible to force parameters at each run, namely:
\enumerate{
\item \strong{fluxes} \cr
(overwriting all or a subset of fluxes defined in \emph{0_ISOBXR_MASTER.xlsx} master file)
\item \strong{isotope fractionation coefficients} \cr
(overwriting all or a subset of coefficients defined in \emph{0_ISOBXR_MASTER.xlsx} master file)
\item \strong{box sizes} \cr
(overwriting all or a subset of box sizes defined in \emph{0_ISOBXR_MASTER.xlsx} master file)
\item \strong{rayleigh isotope distillation}
\item \strong{isotope composition of a source box at initial state}
}
}
\section{Optional outputs}{

\enumerate{
\item If EACH_RUN_DIGEST = TRUE \cr
Creates and fills \strong{\emph{DIGEST}} folder for each run of the composite scenario with
all optional outputs of \code{\link{run_isobxr}} function. \cr
(folder name structure: \strong{\emph{CPS + SERIES_ID + YYY + XXXX + DIGEST}})
\item If to_CPS_DIGEST_CSVs = TRUE \cr
In the \strong{\emph{0_CPS_DIGEST}} folder,
edits csv versions of the whole-composite scenario evD and evS datasets. \cr
(file names structures: \strong{\emph{CPS + SERIES_ID + YYY + evD.csv}}
and \strong{\emph{CPS + SERIES_ID + YYY + evS.csv}})
}
}

\seealso{
Documentation on \code{\link{run_isobxr}}
}
