% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_functions.R
\name{quiet}
\alias{quiet}
\title{Preventing any console prints (for sweepers only)}
\usage{
quiet(x)
}
\arguments{
\item{x}{function to quiet}
}
\value{
No return value, called for side effects
}
\description{
Preventing any console prints (for sweepers only)
}
\keyword{internal}
