% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sweepers.R
\name{sweep_steady}
\alias{sweep_steady}
\title{Sweep the space of two parameters at the final state of a run}
\usage{
sweep_steady(
  workdir,
  SERIES_ID,
  time_units,
  EXPLO_MASTER,
  EXPLO_AXIS_1,
  EXPLO_AXIS_2
)
}
\arguments{
\item{workdir}{Working directory in which the general master file (0_ISOBXR_MASTER.xlsx)
and steady sweep master file (e.g., 0_EXPLO_STEADY_MASTER.xlsx)
are found and where the output files will be stored.
\cr (string of characters)}

\item{SERIES_ID}{Name of the sweep series the run belongs to, that will determine
the folder in which the output files of this sweep run will be stored.
A sweep run number in YYY format will be automatically linked to it and no run will
overwrite a previous sweep run.
\cr (string of characters)}

\item{time_units}{Vector defining the initial time units (identical to time unit
used for fluxes) followed by the time unit used for the graphic output.
\cr (vector of two strings of characters, eg. c("days", "years"))}

\item{EXPLO_MASTER}{Name of the steady sweep master file (e.g., 0_EXPLO_STEADY_MASTER.xlsx),
defining the steady sweep run scenario.
\cr (string of characters)}

\item{EXPLO_AXIS_1}{Set of values of sweeping parameter 1}

\item{EXPLO_AXIS_2}{Set of values of sweeping parameter 2}
}
\value{
If non existing, the fonction creates and stores all outputs
in a SERIES directory located in working directory, automatically numbered.
\cr Directory name structure: 4_STD + SERIES_ID + YYY, YYY being an automically set steady sweep run number between 001 and 999.
\cr No overwriting of previous steady sweep runs is possible.
\enumerate{
\item Calculates the number of runs the sweeping will require depending on the sweeped parameters.
The function then asks the user confirmation to run sweep_steady,
as the run calculation time depends on the number of successive sweeping runs.
\item Creates the set of inputs and outputs for each successive \emph{n} runs, numbered from to 1 to \emph{n} in an XXXX format (see \code{\link{run_isobxr}} documentation).
\cr Single run plots normally edited by \code{\link{run_isobxr}} are here not edited.
\cr Named with the following format: STD + SERIES_ID + YYY + XXXX
\item Archives the steady sweep master file.
\cr (file name structure: 0_STD + SERIES_ID + YYY + EXPLO_STEADY_MASTER.xlsx)
\item Archives the local LOG for the given steady sweep run.
\cr (file name structure: 0_STD + SERIES_ID + YYY + LOG.csv)
\item Stores the evolution of delta values with time in all boxes over the \emph{n} runs that constitute the steady sweep run.
\cr (file name structure: 0_STD + SERIES_ID + YYY + evD.csv)
\item Stores the evolution of box sizes (masses of element X) with time in all boxes over the \emph{n} runs that constitute the steady sweep run
\cr (file name structure: 0_STD + SERIES_ID + YYY + evS.csv)
\item Stores the final state of delta values with time in all boxes over the \emph{n} runs that constitute the steady sweep run.
\cr (file name structure: 0_STD + SERIES_ID + YYY + evD_final.csv)
\item Stores the final state of box sizes (masses of element X) with time in all boxes over the \emph{n} runs that constitute the steady sweep run
\cr (file name structure: 0_STD + SERIES_ID + YYY + evS_final.csv)
}
}
\description{
A function to assess the influence of two parameters (varying
over a range of values) on the final state of a given model.
}
